backend/requirements.txt
altair==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi>=1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.0
contourpy>=1.3,<2
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
fastapi==0.115.12
fasteners==0.19
filelock==3.13.1
Flask==3.0.3
fonttools==4.45.1
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.2
osmnx==1.9.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
parso==0.8.3
passlib==1.7.4
pdbp==1.5.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.41
protobuf==4.25.2
psutil==5.9.6
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyasn1==0.4.8
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pynose==1.5.0
pyodbc==5.1.0
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
pyproj==3.6.1
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.1.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.32.1
requests==2.31.0
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
sbvirtualdisplay==1.3.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.0
selenium==4.18.1
seleniumbase==4.24.5
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.46.1
streamlit==1.30.0
tabcompleter==1.3.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.0
trio==0.24.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.34.0
validators==0.22.0
virtualenv==20.29.1
watchdog==3.0.0
wcwidth==0.2.12
webdriver-manager==4.0.1
Werkzeug==3.0.3
wheel==0.42.0
wsproto==1.2.0
zipp==3.17.0
